fix package structure for TS and Node.js (v6) packaging
