+++ .gitignore
@@ -36,5 +36,6 @@ jspm_packages
 # Optional REPL history
 .node_repl_history
 
-build
+dist
 examples/config.js
+definitions
