--- examples/config.example.js
@@ -1,6 +1,0 @@
-export default {
-  "clientId" : "",
-  "clientSecret" : "",
-  "redirectUri" : "",
-  "accessToken" : ""
-}
