--- examples/fetchUserPlaylists.ts
@@ -1,20 +1,0 @@
-import SpotifyGraphQLClient from '../src/spotify-graphql';
-import config from './config';
-
-SpotifyGraphQLClient.create(config).query(`
-  {
-    user(id: "11879785") {
-      playlists {
-        name
-        tracks {
-          track {
-            id
-            name
-          }
-        }
-      }
-    }
-  }
-`).then(executionResult => {
-  console.log(JSON.stringify(executionResult.data));
-});
