+++ lib/schema.ts
