+++ package.json
@@ -5,11 +5,15 @@
     "type": "git",
     "url": "git://thefrenchhouse/spotify-graphql.git"
   },
+  "main": "dist/index.js",
+  "typings": "definitions/spotify-graphql",
+  "typescript": {
+    "definition": "definitions/spotify-graphql"
+  },
   "description": "GraphQL schema for Spotify WebAPI",
   "engineStrict": true,
   "devDependencies": {
     "@types/es6-shim": "^0.31.32",
-    "@types/immutable": "^3.8.4",
     "@types/node": "~6.0.46",
     "gulp": "^3.9.1",
     "gulp-clean": "^0.3.2",
@@ -17,6 +21,7 @@
     "gulp-typescript": "^3.1.3",
     "immutable": "^3.8.1",
     "jasmine": "^2.5.2",
+    "merge2": "^1.0.2",
     "rimraf": "~2.5.4",
     "run-sequence": "^1.2.2",
     "standard-version": "^3.0.0",
@@ -26,6 +31,7 @@
   },
   "scripts": {
     "test": "gulp test",
+    "compile": "gulp build:es6",
     "release": "standard-version"
   },
   "author": "Charly POLY <hello@charlypoly.com>",
