--- tsconfig.release.json
@@ -1,22 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "sourceMap": false,
-    "declaration": false,
-    "noImplicitAny": false,
-    "rootDir": ".",
-    "outDir": "build",
-    "allowSyntheticDefaultImports": true,
-    "removeComments": true,
-    "jsx": "react",
-    "allowJs": true,
-    "experimentalDecorators": true
-  },
-  "include": [
-    "src/**/*",
-    "examples/**/*",
-    "spec/**/*"
-  ]
-}
